pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evalflow"
version = "0.1.0"
description = "A lightweight, reproducible evaluation pipeline for LLMs with golden datasets and regression/drift monitoring."
readme = "README.md"
requires-python = ">=3.9"
authors = [{name="Joy Ji"}]
license = {text = "MIT"}
dependencies = [
  "pyyaml>=6.0.1",
  "tqdm>=4.66.0",
  "numpy>=1.24.0",
  "jinja2>=3.1.2"
]

[project.scripts]
evalflow = "evalflow.cli:main"
